+++ app/App.tsx
@@ -1,0 +1,20 @@
+import { useRouter } from "expo-router";
+import { Button,Text, View } from "react-native";
+
+export const App = () =>{
+
+    const router = useRouter();
+  return (
+    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+      <Text>Welcome to the Fitness App!</Text>
+      {/* Corrected Link component for navigating to the next screen */}
+
+      <Button
+        title="Get Started"
+        onPress={() => {
+          router.push("/userInput/Gender");
+        }}
+      />
+    </View>
+  );
+}
