+++ app/index.jsx
@@ -1,19 +1,10 @@
-import { useRouter } from "expo-router";
-import { Button, Text, View } from "react-native";
-
+import {App} from './App'
+import { Provider } from 'react-redux';
+import store from './store'
 export default function Home() {
-  const router = useRouter();
   return (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-      <Text>Welcome to the Fitness App!</Text>
-      {/* Corrected Link component for navigating to the next screen */}
-
-      <Button
-        title="Get Started"
-        onPress={() => {
-          router.push("/userInput/Gender");
-        }}
-      />
-    </View>
+    <Provider store={store}>
+      <App />
+   </Provider>
   );
 }
