+++ app/store.ts
@@ -1,0 +1,9 @@
+import { configureStore } from '@reduxjs/toolkit';
+
+const store = configureStore({
+    reducer: {
+    //   counter: counterReducer, // Add reducers here
+    },
+  });
+  
+  export default store;
