+++ app/userInput/Test.tsx
@@ -1,0 +1,10 @@
+import { Link } from "expo-router"
+import { View } from "react-native"
+
+export const Test = () =>{
+    return(
+        <View>
+            <Link href="/(tabs)/Dashboard"></Link>
+        </View>
+    )
+}
